Code reorganization in q3 log parser
