+++ src/quakestats/core/q3toql/entities.py
@@ -1,28 +1,37 @@
-WEAPONS = [
-    "NAILGUN", "BFG", "RAILGUN", "MACHINEGUN", "GRENADE", "SHOTGUN",
-    "LIGHTNING", "PLASMA", "CHAINGUN", "OTHER_WEAPON", "HMG",
-    "PROXMINE", "ROCKET", "GAUNTLET",
-]
+class Q3Data():
+    WEAPONS = [
+        "NAILGUN", "BFG", "RAILGUN", "MACHINEGUN", "GRENADE", "SHOTGUN",
+        "LIGHTNING", "PLASMA", "CHAINGUN", "OTHER_WEAPON", "HMG",
+        "PROXMINE", "ROCKET", "GAUNTLET",
+    ]
 
-MEDALS = [
-    "COMBOKILL", "HEADSHOT", "FIRSTFRAG", "IMPRESSIVE", "CAPTURES",
-    "QUADGOD", "PERFORATED", "EXCELLENT", "ACCURACY", "DEFENDS",
-    "RAMPAGE", "MIDAIR", "REVENGE", "ASSISTS", "HUMILIATION",
-    "PERFECT",
-]
+    MEDALS = [
+        "COMBOKILL", "HEADSHOT", "FIRSTFRAG", "IMPRESSIVE", "CAPTURES",
+        "QUADGOD", "PERFORATED", "EXCELLENT", "ACCURACY", "DEFENDS",
+        "RAMPAGE", "MIDAIR", "REVENGE", "ASSISTS", "HUMILIATION",
+        "PERFECT",
+    ]
 
-PICKUPS = [
-    "HASTE", "FLIGHT", "RED_ARMOR", "BATTLESUIT", "AMMO",
-    "GUARD", "INVIS", "YELLOW_ARMOR", "REGEN", "KAMIKAZE",
-    "DOUBLER", "TELEPORTER", "MEDKIT", "MEGA_HEALTH", "GREEN_ARMOR",
-    "OTHER_HOLDABLE", "INVULNERABILITY", "ARMOR", "HEALTH",
-    "SCOUT", "QUAD", "OTHER_POWERUP", "ARMOR_REGEN", "PORTAL"
-]
+    PICKUPS = [
+        "HASTE", "FLIGHT", "RED_ARMOR", "BATTLESUIT", "AMMO",
+        "GUARD", "INVIS", "YELLOW_ARMOR", "REGEN", "KAMIKAZE",
+        "DOUBLER", "TELEPORTER", "MEDKIT", "MEGA_HEALTH", "GREEN_ARMOR",
+        "OTHER_HOLDABLE", "INVULNERABILITY", "ARMOR", "HEALTH",
+        "SCOUT", "QUAD", "OTHER_POWERUP", "ARMOR_REGEN", "PORTAL"
+    ]
 
-MODS = [
-    'ROCKET', 'ROCKET_SPLASH', 'PLASMA', 'LIGHTNING', 'SHOTGUN',
-    'GAUNTLET', 'PLASMA_SPLASH', 'MACHINEGUN', 'FALLING', 'LAVA',
-    'GRENADE', 'GRENADE_SPLASH', 'BFG', 'BFG_SPLASH', 'TRIGGER_HURT',
-    'RAILGUN', 'TELEFRAG', 'SLIME', 'WATER', 'CRUSH', 'SUICIDE',
-    'SWITCHTEAM'
-]
+    MODS = [
+        'ROCKET', 'ROCKET_SPLASH', 'PLASMA', 'LIGHTNING', 'SHOTGUN',
+        'GAUNTLET', 'PLASMA_SPLASH', 'MACHINEGUN', 'FALLING', 'LAVA',
+        'GRENADE', 'GRENADE_SPLASH', 'BFG', 'BFG_SPLASH', 'TRIGGER_HURT',
+        'RAILGUN', 'TELEFRAG', 'SLIME', 'WATER', 'CRUSH', 'SUICIDE',
+        'SWITCHTEAM'
+    ]
+
+
+class Q3OSPData(Q3Data):
+    pass
+
+
+class Q3EdawnData(Q3Data):
+    pass
